agrosmart/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-tailwind==4.0.1
pillow==11.2.1
sqlparse==0.5.3
